Add Animation Background

+++ Lancelot.Domain/DTOs/Game/Bullets.cs
@@ -2,9 +2,9 @@
 {
     public class Bullets
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
         public bool IsAlive { get; set; }
-        public int Speed { get; set; }
+        public float Speed { get; set; }
     }
 }
